Draft GPU3D renderer modularization
