+++ src/GPU3D.cpp
@@ -273,7 +273,7 @@ u32 RenderNumPolygons;
 u32 FlushRequest;
 u32 FlushAttributes;
 
-
+std::unique_ptr<GPU3D::Renderer3D> CurrentRenderer = {};
 
 bool Init()
 {
@@ -2497,12 +2497,15 @@ void CheckFIFODMA()
 
 void VCount144()
 {
-    if (GPU::Renderer == 0) SoftRenderer::VCount144();
+    CurrentRenderer->VCount144();
 }
 
 void RestartFrame()
 {
-    if (GPU::Renderer == 0) SoftRenderer::SetupRenderThread();
+    if (GPU::Renderer == 0)
+    {
+        reinterpret_cast<SoftRenderer*>(CurrentRenderer.get())->SetupRenderThread();
+    }
 }
 
 
@@ -2597,10 +2600,7 @@ void VBlank()
 
 void VCount215()
 {
-    if (GPU::Renderer == 0) SoftRenderer::RenderFrame();
-#ifdef OGLRENDERER_ENABLED
-    else                    GLRenderer::RenderFrame();
-#endif
+    CurrentRenderer->RenderFrame();
 }
 
 void SetRenderXPos(u16 xpos)
@@ -2614,12 +2614,7 @@ u32 ScrolledLine[256];
 
 u32* GetLine(int line)
 {
-    u32* rawline = NULL;
-
-    if (GPU::Renderer == 0) rawline = SoftRenderer::GetLine(line);
-#ifdef OGLRENDERER_ENABLED
-    else                    rawline = GLRenderer::GetLine(line);
-#endif
+    u32* rawline = CurrentRenderer->GetLine(line);
 
     if (RenderXPos == 0) return rawline;
